Rename the common mod to headers
