+++ src/v2/kuadrant/pipeline/tasks/mod.rs
@@ -1,5 +1,5 @@
 #[allow(dead_code)]
-mod common;
+mod headers;
 mod ratelimit;
 
 use crate::v2::kuadrant::ReqRespCtx;
